refactor(capabilities): migrate the anthropic and gemini caps to the struct-hosted actor macro
Move each cap's `#[actor(singleton)] impl NativeActor` into its runtime
module under `#[runtime]` and lift `#[actor(singleton)]` onto the struct,
per ADR-0123 (#2376). Tests relocate beside the runtime impl so they can
reach the now-runtime-private handler associated fns.
