+++ crates/aether-capabilities/src/anthropic/runtime.rs
@@ -7,17 +7,18 @@
 //! types, gate/reply helpers, and reply assembly through the single
 //! `use runtime::*` glob in the parent.
 
-use super::kinds::{AnthropicError, Message, MessagesSendResult, Role};
+use super::kinds::{AnthropicError, CliSend, Message, MessagesSend, MessagesSendResult, Role};
 use super::{
-    AnthropicAdapter, AnthropicConfig, CliSendResult, CombinedAnthropicAdapter,
-    DisabledAnthropicAdapter, map_adapter_error,
+    AnthropicAdapter, AnthropicCapability, AnthropicConfig, CliSendResult,
+    CombinedAnthropicAdapter, DisabledAnthropicAdapter, map_adapter_error,
 };
-use crate::shared::contentgen::adapter::{AdapterUsage, AnthropicResponse};
+use crate::shared::contentgen::adapter::{AdapterUsage, AnthropicRequest, AnthropicResponse};
 
 pub use crate::shared::contentgen::task_queue::TaskQueue;
 pub use std::sync::Arc;
 
 use aether_actor::OutboundReply;
+use aether_actor::runtime;
 use aether_kinds::Usage;
 
 pub use aether_actor::Manual;
@@ -111,6 +112,137 @@ impl AnthropicCapabilityState {
     }
 }
 
+#[runtime]
+impl NativeActor for AnthropicCapability {
+    /// The runtime state this identity boots into (ADR-0122 split): the
+    /// state-bearing struct holding the adapter + the rate-limit queue.
+    type State = AnthropicCapabilityState;
+
+    type Config = AnthropicConfig;
+
+    /// ADR-0050 + ADR-0074 Phase 5 chassis-owned mailbox.
+    const NAMESPACE: &'static str = "aether.anthropic";
+
+    /// Build the adapter from the resolved config and capture the
+    /// mailer + own mailbox so the spawn-and-die helper can land
+    /// loopback result mails. The adapter is built immediately so a
+    /// key-absent boot still loads (replying Unauthorized) rather
+    /// than warn-dropping.
+    fn init(
+        config: AnthropicConfig,
+        _ctx: &mut NativeInitCtx<'_>,
+    ) -> Result<AnthropicCapabilityState, BootError> {
+        Ok(AnthropicCapabilityState {
+            adapter: build_adapter(&config),
+            tasks: TaskQueue::new(config.max_in_flight),
+        })
+    }
+
+    /// Run a Messages-API completion off the dispatcher thread.
+    ///
+    /// # Agent
+    /// Reply: `MessagesSendResult`. Validates `model` against the
+    /// supported table synchronously (`UnknownModel` on a miss),
+    /// then dispatches the blocking HTTPS call on an ephemeral
+    /// thread; the reply lands when the call returns.
+    #[handler::manual]
+    fn on_messages_send(
+        state: &mut Self::State,
+        ctx: &mut NativeCtx<'_, Manual>,
+        mail: MessagesSend,
+    ) {
+        let request_id = mail.request_id;
+        if !state.gate_model(ctx, SendPath::Messages, request_id, &mail.model) {
+            return;
+        }
+        let req = AnthropicRequest {
+            prompt: flatten_prompt(&mail.messages),
+            model: mail.model,
+            system: mail.system,
+            max_tokens: mail.max_tokens,
+            temperature: mail.temperature,
+        };
+        let adapter = Arc::clone(&state.adapter);
+        state.tasks.submit(ctx, move || {
+            let result = adapter.messages_send(req);
+            messages_reply(request_id, result)
+        });
+    }
+
+    /// Run a `claude`-subprocess completion off the dispatcher
+    /// thread.
+    ///
+    /// # Agent
+    /// Reply: `CliSendResult`. Replies `Err { CliNotFound }` when
+    /// `claude` isn't on PATH. The CLI uses the user's subscription,
+    /// so it works even when `ANTHROPIC_API_KEY` is unset. The
+    /// `claude` binary exposes no `--max-tokens` / `--temperature`
+    /// flag, so setting either replies `Err { ParamNotSupported }`
+    /// synchronously (no dispatch) rather than silently dropping it —
+    /// route sampling knobs through `aether.anthropic.messages.send`.
+    #[handler::manual]
+    fn on_cli_send(state: &mut Self::State, ctx: &mut NativeCtx<'_, Manual>, mail: CliSend) {
+        // The `claude` CLI has no flag for either knob; reject when
+        // set instead of silently dropping (the outcome to avoid —
+        // `feedback_explicit_nulls_over_absent_fields`).
+        let mut unsupported = Vec::new();
+        if mail.max_tokens.is_some() {
+            unsupported.push("max_tokens");
+        }
+        if mail.temperature.is_some() {
+            unsupported.push("temperature");
+        }
+        if !unsupported.is_empty() {
+            let error = AnthropicError::ParamNotSupported {
+                param: unsupported.join(", "),
+                reason: "the claude CLI has no flag for this; use aether.anthropic.messages.send"
+                    .to_string(),
+            };
+            AnthropicCapabilityState::reply_err(ctx, SendPath::Cli, mail.request_id, error);
+            return;
+        }
+        let request_id = mail.request_id;
+        // CLI passes the model through to `claude` (no gate), so no
+        // `gate_model` call here.
+        let req = AnthropicRequest {
+            prompt: flatten_prompt(&mail.messages),
+            model: mail.model,
+            system: mail.system,
+            max_tokens: mail.max_tokens,
+            temperature: mail.temperature,
+        };
+        let adapter = Arc::clone(&state.adapter);
+        state.tasks.submit(ctx, move || {
+            let result = adapter.cli_send(req);
+            cli_reply(request_id, result)
+        });
+    }
+
+    /// ADR-0093 completion for a finished Messages call: re-reply the
+    /// worker's result to the original caller (drops the hold), then
+    /// free the in-flight slot (draining the next pending request).
+    #[handler(task)]
+    fn on_messages_done(
+        state: &mut Self::State,
+        ctx: &mut NativeCtx<'_>,
+        done: TaskDone<MessagesSendResult>,
+    ) {
+        done.resolve(ctx);
+        state.tasks.on_complete(ctx);
+    }
+
+    /// ADR-0093 completion for a finished CLI call.
+    #[handler(task)]
+    fn on_cli_done(
+        state: &mut Self::State,
+        ctx: &mut NativeCtx<'_>,
+        done: TaskDone<CliSendResult>,
+    ) {
+        done.resolve(ctx);
+        state.tasks.on_complete(ctx);
+    }
+}
+
 pub fn build_adapter(config: &AnthropicConfig) -> Arc<dyn AnthropicAdapter> {
     if config.disabled {
         tracing::info!(
@@ -194,5 +326,419 @@ pub fn cli_reply(request_id: u64, result: Result<AnthropicResponse, String>) -> 
             request_id,
             error: map_adapter_error(&raw),
         },
+    }
+}
+
+#[cfg(all(test, feature = "runtime"))]
+mod tests {
+    use super::AnthropicCapabilityState;
+    use crate::anthropic::{
+        AnthropicAdapter, AnthropicConfig, ClaudeCliAdapter, DisabledAnthropicAdapter,
+    };
+    use crate::anthropic::{
+        AnthropicCapability, AnthropicError, CliSend, CliSendResult, Message, MessagesSend,
+        MessagesSendResult, Role,
+    };
+    use crate::shared::contentgen::adapter::{
+        AnthropicRequest, AnthropicResponse, StubAnthropicAdapter,
+    };
+    use crate::test_chassis::{
+        TestChassis, decode_session_reply, drive_task_completion, fresh_substrate,
+        test_mailer_and_rx,
+    };
+    use aether_actor::Addressable;
+    use aether_data::{Kind, MailboxId, SessionToken, Source, SourceAddr, Uuid};
+    use aether_substrate::actor::native::binding::NativeBinding;
+    use aether_substrate::actor::native::ctx::NativeCtx;
+    use aether_substrate::chassis::builder::Builder;
+    use aether_substrate::mail::mailer::Mailer;
+    use aether_substrate::mail::outbound::EgressEvent;
+    use serde::de::DeserializeOwned;
+    use std::sync::Arc;
+    use std::sync::mpsc::Receiver;
+    use std::time::Duration;
+
+    fn session_sender() -> Source {
+        Source::to(SourceAddr::Session(SessionToken(Uuid::nil())))
+    }
+
+    fn user_msg(text: &str) -> Vec<Message> {
+        vec![Message {
+            role: Role::User,
+            content: text.to_string(),
+        }]
+    }
+
+    /// Thin alias over the shared `decode_session_reply` so call
+    /// sites stay terse.
+    fn decode_reply<K: Kind + DeserializeOwned>(rx: &Receiver<EgressEvent>) -> K {
+        decode_session_reply(rx)
+    }
+
+    /// Adapter that records the prompt it saw and returns canned text.
+    struct RecordingStub {
+        inner: StubAnthropicAdapter,
+    }
+
+    impl AnthropicAdapter for RecordingStub {
+        fn messages_send(&self, req: AnthropicRequest) -> Result<AnthropicResponse, String> {
+            self.inner.messages_send(req)
+        }
+        fn cli_send(&self, req: AnthropicRequest) -> Result<AnthropicResponse, String> {
+            self.inner.cli_send(req)
+        }
+        fn supported_models(&self) -> Vec<String> {
+            vec!["claude-test".to_string()]
+        }
+    }
+
+    /// Boot the cap against a default (key-absent) config and confirm
+    /// the mailbox registers.
+    #[test]
+    fn capability_boots_and_registers_mailbox() {
+        let (registry, mailer) = fresh_substrate();
+        let chassis = Builder::<TestChassis>::new(Arc::clone(&registry), Arc::clone(&mailer))
+            .with_actor::<AnthropicCapability>(AnthropicConfig::default())
+            .build_passive()
+            .expect("anthropic capability boots");
+        assert!(
+            registry.lookup(AnthropicCapability::NAMESPACE).is_some(),
+            "anthropic mailbox registered"
+        );
+        drop(chassis);
+    }
+
+    /// Drive a stub Messages request end-to-end through the ADR-0093
+    /// dispatch primitive: the cap submits to the `TaskQueue`, the real
+    /// worker runs the stub call, pushes a completion wake, and the
+    /// cap's `#[handler(task)]` re-replies the `Ok` to the caller.
+    #[test]
+    fn anthropic_stub_messages() {
+        let (mailer, rx) = test_mailer_and_rx();
+        let cap_mailbox = MailboxId(0);
+        let mut state = AnthropicCapabilityState::from_parts(
+            Arc::new(RecordingStub {
+                inner: StubAnthropicAdapter::default(),
+            }),
+            4,
+        );
+        let transport = Arc::new(NativeBinding::new_for_test(
+            Arc::clone(&mailer),
+            cap_mailbox,
+        ));
+        let mut ctx = NativeCtx::new_dispatching(
+            &transport,
+            session_sender(),
+            aether_data::MailId::NONE,
+            aether_data::MailId::NONE,
+        );
+        AnthropicCapability::on_messages_send(
+            &mut state,
+            &mut ctx,
+            MessagesSend {
+                request_id: 7,
+                model: "claude-test".to_string(),
+                messages: user_msg("hi"),
+                max_tokens: Some(8),
+                temperature: None,
+                system: None,
+            },
+        );
+        // The worker runs the stub call and pushes the completion wake;
+        // route it through the cap's task handler.
+        drive_task_completion::<AnthropicCapability>(&mut state, &transport, &rx);
+        match decode_reply::<MessagesSendResult>(&rx) {
+            MessagesSendResult::Ok {
+                request_id, text, ..
+            } => {
+                assert_eq!(request_id, 7);
+                assert_eq!(text, "stub completion");
+            }
+            other @ MessagesSendResult::Err { .. } => panic!("expected Ok, got {other:?}"),
+        }
+    }
+
+    /// Unknown model errors synchronously, before any dispatch.
+    #[test]
+    fn anthropic_unknown_model_errors() {
+        let (mailer, rx) = test_mailer_and_rx();
+        let cap_mailbox = MailboxId(0);
+        let mut state = AnthropicCapabilityState::from_parts(
+            Arc::new(RecordingStub {
+                inner: StubAnthropicAdapter::default(),
+            }),
+            4,
+        );
+        let transport = Arc::new(NativeBinding::new_for_test(
+            Arc::clone(&mailer),
+            cap_mailbox,
+        ));
+        let mut ctx = NativeCtx::new_dispatching(
+            &transport,
+            session_sender(),
+            aether_data::MailId::NONE,
+            aether_data::MailId::NONE,
+        );
+        AnthropicCapability::on_messages_send(
+            &mut state,
+            &mut ctx,
+            MessagesSend {
+                request_id: 3,
+                model: "claude-bogus".to_string(),
+                messages: user_msg("hi"),
+                max_tokens: None,
+                temperature: None,
+                system: None,
+            },
+        );
+        match decode_reply::<MessagesSendResult>(&rx) {
+            MessagesSendResult::Err {
+                request_id,
+                error: AnthropicError::UnknownModel { model, supported },
+            } => {
+                assert_eq!(request_id, 3);
+                assert_eq!(model, "claude-bogus");
+                assert!(supported.contains(&"claude-test".to_string()));
+            }
+            other => panic!("expected UnknownModel, got {other:?}"),
+        }
+        // No in-flight work was spawned — the synchronous error path
+        // never touched the dispatch helper.
+        assert_eq!(cap_in_flight(&state), 0);
+    }
+
+    /// Boot a cap against the recording stub and fire a `CliSend`
+    /// carrying the given knobs at `on_cli_send`, returning the state so
+    /// the caller can assert `test_in_flight()`. The reply lands on
+    /// the `mailer`'s loopback rx (held separately by the caller).
+    fn cli_send_with(
+        mailer: &Arc<Mailer>,
+        max_tokens: Option<u32>,
+        temperature: Option<f32>,
+    ) -> AnthropicCapabilityState {
+        let cap_mailbox = MailboxId(0);
+        let mut state = AnthropicCapabilityState::from_parts(
+            Arc::new(RecordingStub {
+                inner: StubAnthropicAdapter::default(),
+            }),
+            4,
+        );
+        let transport = Arc::new(NativeBinding::new_for_test(Arc::clone(mailer), cap_mailbox));
+        let mut ctx = NativeCtx::new_dispatching(
+            &transport,
+            session_sender(),
+            aether_data::MailId::NONE,
+            aether_data::MailId::NONE,
+        );
+        AnthropicCapability::on_cli_send(
+            &mut state,
+            &mut ctx,
+            CliSend {
+                request_id: 11,
+                model: "claude-test".to_string(),
+                messages: user_msg("hi"),
+                max_tokens,
+                temperature,
+                system: None,
+            },
+        );
+        state
+    }
+
+    /// `on_cli_send` with `max_tokens` set replies
+    /// `Err { ParamNotSupported }` synchronously and spawns no work —
+    /// the `claude` CLI has no flag to honor it.
+    #[test]
+    fn anthropic_cli_rejects_max_tokens() {
+        let (mailer, rx) = test_mailer_and_rx();
+        let state = cli_send_with(&mailer, Some(256), None);
+        match decode_reply::<CliSendResult>(&rx) {
+            CliSendResult::Err {
+                request_id,
+                error: AnthropicError::ParamNotSupported { param, reason },
+            } => {
+                assert_eq!(request_id, 11);
+                assert!(param.contains("max_tokens"), "param was {param:?}");
+                assert!(reason.contains("messages.send"), "reason was {reason:?}");
+            }
+            other => panic!("expected ParamNotSupported, got {other:?}"),
+        }
+        // Synchronous error path never touched the dispatch helper.
+        assert_eq!(cap_in_flight(&state), 0);
+    }
+
+    /// `on_cli_send` with `temperature` set replies
+    /// `Err { ParamNotSupported }` synchronously and spawns no work.
+    #[test]
+    fn anthropic_cli_rejects_temperature() {
+        let (mailer, rx) = test_mailer_and_rx();
+        let state = cli_send_with(&mailer, None, Some(0.7));
+        match decode_reply::<CliSendResult>(&rx) {
+            CliSendResult::Err {
+                request_id,
+                error: AnthropicError::ParamNotSupported { param, .. },
+            } => {
+                assert_eq!(request_id, 11);
+                assert!(param.contains("temperature"), "param was {param:?}");
+            }
+            other => panic!("expected ParamNotSupported, got {other:?}"),
+        }
+        assert_eq!(cap_in_flight(&state), 0);
+    }
+
+    /// A `CliSend` with both knobs `None` dispatches normally — the
+    /// synchronous reject path is skipped and work is spawned.
+    #[test]
+    fn anthropic_cli_no_params_dispatches() {
+        let (mailer, _rx) = test_mailer_and_rx();
+        let state = cli_send_with(&mailer, None, None);
+        assert_eq!(
+            cap_in_flight(&state),
+            1,
+            "a param-free CliSend should dispatch one in-flight call"
+        );
+    }
+
+    /// CLI send with a missing `claude` binary replies
+    /// `Err { CliNotFound }` — a graceful skip, not a hang.
+    #[test]
+    fn anthropic_cli_skips_when_no_claude_on_path() {
+        use crate::anthropic::error::UNAUTHORIZED_SENTINEL;
+        // A disabled adapter routes CLI through the real subprocess
+        // backend; pointing it at a missing binary exercises the
+        // CliNotFound path without depending on the host's `claude`.
+        struct MissingCliAdapter {
+            cli: ClaudeCliAdapter,
+        }
+        impl AnthropicAdapter for MissingCliAdapter {
+            fn messages_send(&self, _req: AnthropicRequest) -> Result<AnthropicResponse, String> {
+                Err(UNAUTHORIZED_SENTINEL.to_string())
+            }
+            fn cli_send(&self, req: AnthropicRequest) -> Result<AnthropicResponse, String> {
+                self.cli.cli_send(&req)
+            }
+        }
+
+        let (mailer, rx) = test_mailer_and_rx();
+        let cap_mailbox = MailboxId(0);
+        let mut state = AnthropicCapabilityState::from_parts(
+            Arc::new(MissingCliAdapter {
+                cli: ClaudeCliAdapter::new(
+                    "aether-nonexistent-claude-binary-xyzzy".to_string(),
+                    Duration::from_secs(30),
+                ),
+            }),
+            4,
+        );
+        let transport = Arc::new(NativeBinding::new_for_test(
+            Arc::clone(&mailer),
+            cap_mailbox,
+        ));
+        let mut ctx = NativeCtx::new_dispatching(
+            &transport,
+            session_sender(),
+            aether_data::MailId::NONE,
+            aether_data::MailId::NONE,
+        );
+        AnthropicCapability::on_cli_send(
+            &mut state,
+            &mut ctx,
+            CliSend {
+                request_id: 5,
+                model: "claude-test".to_string(),
+                messages: user_msg("hi"),
+                max_tokens: None,
+                temperature: None,
+                system: None,
+            },
+        );
+        // The CLI backend runs on the real worker against a missing
+        // binary, yielding CliNotFound; route the completion through the
+        // cap's task handler.
+        drive_task_completion::<AnthropicCapability>(&mut state, &transport, &rx);
+        match decode_reply::<CliSendResult>(&rx) {
+            CliSendResult::Err {
+                request_id,
+                error: AnthropicError::CliNotFound,
+            } => {
+                assert_eq!(request_id, 5);
+            }
+            other => panic!("expected CliNotFound, got {other:?}"),
+        }
+    }
+
+    /// Disabled adapter replies `Unauthorized` to a Messages request.
+    #[test]
+    fn anthropic_disabled_messages_replies_unauthorized() {
+        let (mailer, rx) = test_mailer_and_rx();
+        let cap_mailbox = MailboxId(0);
+        let mut state =
+            AnthropicCapabilityState::from_parts(Arc::new(DisabledAnthropicAdapter::default()), 4);
+        let transport = Arc::new(NativeBinding::new_for_test(
+            Arc::clone(&mailer),
+            cap_mailbox,
+        ));
+        let mut ctx = NativeCtx::new_dispatching(
+            &transport,
+            session_sender(),
+            aether_data::MailId::NONE,
+            aether_data::MailId::NONE,
+        );
+        AnthropicCapability::on_messages_send(
+            &mut state,
+            &mut ctx,
+            MessagesSend {
+                request_id: 9,
+                model: "claude-anything".to_string(),
+                messages: user_msg("hi"),
+                max_tokens: None,
+                temperature: None,
+                system: None,
+            },
+        );
+        // Disabled adapter has an empty supported-models table, so the
+        // model gate is skipped and the request dispatches; the worker
+        // produces the Unauthorized result. Route the completion through
+        // the cap's task handler.
+        drive_task_completion::<AnthropicCapability>(&mut state, &transport, &rx);
+        match decode_reply::<MessagesSendResult>(&rx) {
+            MessagesSendResult::Err {
+                request_id,
+                error: AnthropicError::Unauthorized,
+            } => assert_eq!(request_id, 9),
+            other => panic!("expected Unauthorized, got {other:?}"),
+        }
+    }
+
+    /// Real-API smoke. Hits the live Messages API with a tiny
+    /// 5-`max_tokens` request — ignored by default so CI stays
+    /// zero-cost; run with `ANTHROPIC_API_KEY` set.
+    #[test]
+    #[ignore = "needs ANTHROPIC_API_KEY"]
+    fn anthropic_api_smoke() {
+        use crate::anthropic::UreqAnthropicAdapter;
+        use crate::shared::contentgen::adapter::AnthropicRequest;
+        use std::env;
+        // Test-only: the live-API smoke reads an external credential
+        // (ANTHROPIC_API_KEY), not cap config; gated `#[ignore]`.
+        #[allow(clippy::disallowed_methods)]
+        let key = env::var("ANTHROPIC_API_KEY").expect("ANTHROPIC_API_KEY set for smoke");
+        let adapter = UreqAnthropicAdapter::new(key, Duration::from_secs(30));
+        let resp = adapter
+            .messages_send(&AnthropicRequest {
+                model: "claude-haiku-4-5-20251001".to_string(),
+                prompt: "say hi".to_string(),
+                system: None,
+                max_tokens: Some(5),
+                temperature: None,
+            })
+            .expect("live messages request succeeds");
+        assert!(!resp.text.is_empty());
+    }
+
+    // White-box accessor for the queue's in-flight count; the state's
+    // `tasks` field is private, so tests read it through this shim.
+    fn cap_in_flight(state: &AnthropicCapabilityState) -> usize {
+        state.test_in_flight()
     }
 }
